scripts/verify/requirements.txt
kafka-python>=2.0.2
redis>=4.5.0
pandas>=1.5.0
tqdm>=4.65.0
